+++ lib/angular_xss/action_view.rb
@@ -1,5 +1,4 @@
 ActionView::Template.class_eval do
-
   protected
 
   def compile_with_angular_xss(*args, &block)
@@ -10,5 +9,4 @@ ActionView::Template.class_eval do
 
   alias_method :compile_without_angular_xss, :compile
   alias_method :compile, :compile_with_angular_xss
-
 end